# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TwinBlueprint Server API",
    "description": "REST API for TwinBlueprint — leads, bids, projects, campaigns, analytics, and admin auth.",
    "version": "3.0.0",
    "contact": {
      "name": "TwinBlueprint",
      "email": "twinblueprints@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token. Can also be sent via cookie named 'token'."
      }
    },
    "schemas": {
...
      "CreateLeadInput": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "minLength": 2,
            "example": "Jane Smith"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane@acme.com"
          },
          "company": {
            "type": "string",
            "example": "Acme Inc."
          },
          "job_title": {
            "type": "string",
            "example": "CTO"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "applications": {
            "type": "integer",
            "minimum": 0,
            "default": 0,
            "example": 3
          },
          "score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "default": 0,
            "example": 75
          },
          "region": {
            "type": "string",
            "nullable": true,
            "maxLength": 100,
            "description": "Granular regional classification (e.g. Middle East (GCC) or Latin America), populated by the AI/enrichment pipeline. Free text rather than an enum; NULL means enrichment could not confidently determine it.",
            "example": "UK & Ireland"
          },
          "country": {
            "type": "string",
            "nullable": true,
            "maxLength": 100,
            "description": "Explicit recorded country, populated by the AI/enrichment pipeline. Free text rather than an enum so any enriched value (e.g. Saudi Arabia, Germany) is accepted; NULL means enrichment could not confidently determine it.",
            "example": "United Kingdom"
          },
          "project": {
            "type": "string",
            "nullable": true,
            "example": "LAX Terminal Modernization"
          },
          "project_size": {
            "type": "string",
            "nullable": true,
            "example": "$2.1B"
          },
          "phase": {
            "type": "string",
            "nullable": true,
            "enum": [
              "Discovery",
              "Bid",
              "In-flight"
            ]
          },
          "lead_status": {
            "type": "string",
            "nullable": true,
            "enum": [
              "New",
              "Identified",
              "Bidding",
              "Inflight",
              "Closed"
            ]
          },
          "archived": {
            "type": "boolean",
            "example": false
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "contacted",
              "qualified",
              "proposal",
              "negotiation",
              "won",
              "lost"
            ],
            "default": "new"
          },
          "assigned_to": {
            "type": "integer",
            "nullable": true
          },
          "temperature": {
            "type": "string",
            "enum": [
              "hot",
              "warm",
              "cool"
            ],
            "example": "hot"
          },
          "application_tools": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Revit",
              "Navisworks"
            ]
          },
          "industry": {
            "type": "string",
            "nullable": true,
            "enum": [
              "Construction",
              "Architecture",
              "Urban Development",
              "Infrastructure",
              "Engineering",
              "Other"
            ],
            "example": "Construction"
          },
          "project_value": {
            "type": "number",
            "nullable": true,
            "description": "Numeric lead deal size in currency units. NULL means unknown; lower bound of a range when parsed from legacy project_size."
          },
          "currency": {
            "type": "string",
            "nullable": true,
            "enum": [
              "USD",
              "EUR",
              "GBP",
              "NGN"
            ],
            "description": "Explicit source ISO 4217 currency. NULL means unknown; never backfilled from geography."
          },
          "region_group": {
            "type": "string",
            "nullable": true,
            "maxLength": 100,
            "description": "Broad global cluster (Americas/EMEA/APAC/Other), populated by the AI/enrichment pipeline. Free text rather than an enum; NULL means enrichment could not confidently determine it.",
            "example": "EMEA"
          },
          "state": {
            "type": "string",
            "nullable": true,
            "maxLength": 100,
            "description": "Administrative subdivision (state, province, department, district) of the lead country, populated by the AI/enrichment pipeline. Free text; NULL means enrichment could not confidently determine it.",
            "example": "Texas"
          }
        },
        "required": [
          "full_name",
          "email"
        ]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/leads": {
      "post": {
        "tags": [
          "Leads"
        ],
        "summary": "Create a lead",
        "description": "Creates a new lead. Status defaults to 'new'. Triggers an async email notification.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeadInput"
              },
              "example": {
                "full_name": "Michael Okafor",
                "email": "m.okafor@aecom.com",
                "company": "AECOM",
                "industry": "Construction",
                "region_group": "EMEA",
                "region": "UK & Ireland",
                "state": "Greater London",
                "country": "United Kingdom",
                "project": "LAX Terminal Modernization",
                "phase": "Bid",
                "lead_status": "Bidding",
                "status": "qualified",
                "applications": 9,
                "application_tools": [
                  "Revit",
                  "Navisworks"
                ],
                "score": 87,
                "temperature": "hot",
                "archived": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lead created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "lead": {
                          "$ref": "#/components/schemas/Lead"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "A lead with this email already exists"
          }
        }
      },
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "List all leads",
        "description": "Lists leads with filtering, pagination, and sorting. Sort by company, industry, project, full_name, region, project_size, phase, status, lead_status, score, created_at, or updated_at. `archived` is the archive source of truth.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "industry",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project_size",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phase",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lead_status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "temperature",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_score",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "archived",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "leads": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Lead"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/pipeline": {
      "get": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Get pipeline data",
        "description": "Returns stages determined by lead status. Each stage includes all bids linked to its leads, regardless of bid status, with lead_id, nullable numeric value, currency, and suppliers. A stage currency is reported only when every linked bid shares one source ISO currency; mixed or unknown currencies report null. The separate active_bids list and summary remain limited to active bids. inflight_projects contains full saved projects with status In-flight; Active, On Hold, Completed and Cancelled are excluded. Each active bid includes supplier_details from explicitly linked supplier profiles. suppliers retains legacy names plus linked profile names. Unlinked names are not converted into fabricated profiles. Apply supplier-migration.sql before deploying.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Pipeline data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PipelineResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/bids": {
      "post": {
        "tags": [
          "Bids"
        ],
        "summary": "Create a bid",
        "description": "Admin only. Create a bid using project, client, phase, and deadline. Supply lead_id to link an existing lead without changing its status or project_size. value defaults to null and suppliers to []. Duplicate lead links return 409; update the existing bid instead.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBidInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bid created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "bid": {
                          "$ref": "#/components/schemas/Bid"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Bid or linked lead not found."
          },
          "409": {
            "description": "The lead already has a bid. Update that bid instead."
          }
        }
      },
      "get": {
        "tags": [
          "Bids"
        ],
        "summary": "List all bids",
        "description": "Supports pagination.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "bids": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Bid"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create a project",
        "description": "Admin only. Requires project, client, start_date, end_date. Saves progress (0–100), supplier names, uses_3d, competitor, issue, bid_id and status. Status defaults to Active; only In-flight appears in pipeline inflight_projects.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "project": {
                          "$ref": "#/components/schemas/Project"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "List all projects",
        "description": "Supports pagination.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "projects": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Project"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/campaigns": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Create a campaign",
        "description": "Admin only.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "campaign": {
                          "$ref": "#/components/schemas/Campaign"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "List all campaigns",
        "description": "Supports pagination.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "campaigns": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Campaign"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/demo": {
      "post": {
        "tags": [
          "Demo"
        ],
        "summary": "Submit a demo request",
        "description": "Public endpoint. Rate limited to 5 requests per hour per IP. Creates a lead record and triggers email notification.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Demo request submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Lead"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/regions": {
      "get": {
        "tags": [
          "Regions"
        ],
        "summary": "List supported lead regions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Region options",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "example": {
                        "regions": [
                          "North America",
                          "Latin America",
                          "UK & Ireland",
                          "DACH / Northern Europe",
                          "Middle East (GCC)",
                          "Africa",
                          "Southern Europe"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
...
  }
}